+++ src/arch/src/lib.rs
@@ -1,2 +1,7 @@
 #![feature(asm_const)]
+#![feature(naked_functions)]
+#![feature(generator_trait)]
 #![no_std]
+
+#[cfg(feature = "riscv64")]
+pub mod riscv64;
