+++ src/arch/src/riscv64/mod.rs
@@ -1,0 +1,7 @@
+pub mod sbi {
+    pub mod execute;
+    pub mod feature;
+    pub mod hart_csr_utils;
+    pub mod info;
+    pub mod runtime;
+}
