+++ src/arch/src/riscv64/sbi/feature/emulate_rdtime.rs
@@ -1,7 +1,7 @@
-use crate::runtime::SupervisorContext;
+use super::super::runtime::SupervisorContext;
 use core::arch::asm;
+use log::println;
 use riscv::register::cycle;
-use rustsbi::println;
 
 const DEBUG_THIS: bool = false;
