+++ src/arch/src/riscv64/sbi/feature/sfence_vma.rs
@@ -1,4 +1,4 @@
-use crate::runtime::SupervisorContext;
+use super::super::runtime::SupervisorContext;
 use core::arch::asm;
 use riscv::register::{mstatus, satp};
