+++ src/mainboard/sunxi/nezha/bt0/Cargo.toml
@@ -8,8 +8,6 @@ authors = [
 ]
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 embedded-hal = "=1.0.0-alpha.9"
 embedded-hal-nb = "=1.0.0-alpha.1"
