+++ src/mainboard/sunxi/nezha/main/Cargo.toml
@@ -3,31 +3,22 @@ name = "oreboot-nezha-main"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 bitflags = "1"
 embedded-hal = "=1.0.0-alpha.9"
 embedded-hal-nb = "=1.0.0-alpha.1"
 lazy_static = { version = "1", features = ["spin_no_std"] }
 nb = "1"
-# rustsbi = { path = "../rustsbi", features =["legacy"] }
-rustsbi = { version = "0.3.1", features = ["legacy"] }
+riscv = { version = "0.10.0", features = ["critical-section-single-hart"] }
+rustsbi = "0.3.2"
 sbi-spec = "0.0.4"
-# todo: a new package oreboot-console
-spin = "0.9" # todo: remove when we have a console package
+spin = "0.9"
 vcell = "0.1.3"
 
-[dependencies.oreboot_compression]
-path = "../compression"
-
-[dependencies.oreboot-soc]
-path = "../../../../soc"
-features = ["sunxi_d1"]
-
-[dependencies.riscv]
-version = "0.10.0"
-features = ["critical-section-single-hart"]
+log = { path = "../../../../lib/log" }
+oreboot-arch = { path = "../../../../arch", features = ["riscv64"] }
+oreboot-soc = { path = "../../../../soc", features = ["sunxi_d1"] }
+oreboot_compression = { path = "../compression" }
 
 [features]
 default = []
